Updated education route to allow read or change only when cv is published or you're owner. Path is now nested.

+++ src/routes/cv/education/education.service.js
@@ -1,4 +1,5 @@
 const db = require('src/_helpers/db');
+const cvService = require('../cv.service');
 
 module.exports = {
     getAll,
@@ -8,24 +9,33 @@ module.exports = {
     delete: _delete
 };
 
-async function getAll() {
-    return await db.CVEducation.findAll();
+async function getAll(userId, cvId) {
+    await cvService.validateOwnership(cvId, userId);
+
+    return await db.CVEducation.findAll({
+        where: { cvId }
+    });
 }
 
-async function getById(id) {
-    const education = await getEducation(id);
+async function getById(userId, id, cvId) {
+    const education = await getEducation(userId, id, cvId);
 
     return education;
 }
 
-async function set(params) {
+async function set(userId, cvId, params) {
+    await cvService.validateOwnership(cvId, userId);
+
     const { school, degree, location, details } = params;
     const schoolString = await db.Strings.create(school);
     const degreeString = await db.Strings.create(degree);
     const locationString = await db.Strings.create(location);
     const detailsString = await db.Strings.create(details);
 
-    const education = await db.CVEducation.create(params);
+    const education = await db.CVEducation.create({
+        ...params,
+        cvId
+    });
 
     education.setSchool(schoolString);
     education.setDegree(degreeString);
@@ -33,9 +43,11 @@ async function set(params) {
     education.setDetails(detailsString);
 }
 
-async function update(id, params) {
+async function update(userId, id, cvId, params) {
+    await cvService.validateOwnership(cvId, userId);
+
     const { school, degree, location, details } = params;
-    const education = await getEducation(id);
+    const education = await getEducation(userId, id, cvId);
     const { schoolId, degreeId, locationId, detailsId } = education.get({ plain: true });
     await db.Strings.update(school, { where: { id: schoolId }});
     await db.Strings.update(degree, { where: { id: degreeId }});
@@ -46,14 +58,18 @@ async function update(id, params) {
     await education.save();
 }
 
-async function getEducation(id) {
-    const education = await db.CVEducation.findByPk(id);
+async function getEducation(userId, id, cvId) {
+    await cvService.validateOwnership(cvId, userId);
+
+    const education = await db.CVEducation.findOne({ where: { id, cvId }});
     if (!education) throw 'CVEducation not found';
     return education;
 }
 
-async function _delete(id) {
-    const education = await getEducation(id);
+async function _delete(userId, id, cvId) {
+    await cvService.validateOwnership(cvId, userId);
+
+    const education = await getEducation(userId, id, cvId);
     const { schoolId, degreeId, locationId, detailsId } = education.get({ plain: true });
     await db.Strings.destroy({ where: { id: schoolId }});
     await db.Strings.destroy({ where: { id: degreeId }});
